requirements.txt
pyyaml
beautifulsoup4
urllib3
colorama
playwright